fix: move getClientXY to util/dom
