+++ packages/components/color-picker/src/components/alpha-slider.vue
@@ -28,7 +28,7 @@ import {
   getCurrentInstance,
   shallowRef,
 } from 'vue'
-import { getClientXY } from '@element-plus/utils/util'
+import { getClientXY } from '@element-plus/utils/dom'
 import draggable from '../draggable'
 
 import type { PropType } from 'vue'
