+++ packages/components/color-picker/src/components/sv-panel.vue
@@ -28,7 +28,7 @@ import {
   getCurrentInstance,
   onMounted,
 } from 'vue'
-import { getClientXY } from '@element-plus/utils/util'
+import { getClientXY } from '@element-plus/utils/dom'
 import draggable from '../draggable'
 
 import type { PropType } from 'vue'
