+++ packages/utils/dom.ts
@@ -234,3 +234,22 @@ export const getOffsetTopDistance = (
 }
 
 export const stop = (e: Event) => e.stopPropagation()
+
+export const getClientXY = (event: MouseEvent | TouchEvent) => {
+  let clientX: number
+  let clientY: number
+  if (event.type === 'touchend') {
+    clientY = (event as TouchEvent).changedTouches[0].clientY
+    clientX = (event as TouchEvent).changedTouches[0].clientX
+  } else if (event.type.startsWith('touch')) {
+    clientY = (event as TouchEvent).touches[0].clientY
+    clientX = (event as TouchEvent).touches[0].clientX
+  } else {
+    clientY = (event as MouseEvent).clientY
+    clientX = (event as MouseEvent).clientX
+  }
+  return {
+    clientX,
+    clientY,
+  }
+}
