+++ packages/utils/util.ts
@@ -248,22 +248,3 @@ export const refAttacher = <T extends HTMLElement | ComponentPublicInstance>(
     ref.value = val
   }
 }
-
-export const getClientXY = (event: MouseEvent | TouchEvent) => {
-  let clientX: number
-  let clientY: number
-  if (event.type === 'touchend') {
-    clientY = (event as TouchEvent).changedTouches[0].clientY
-    clientX = (event as TouchEvent).changedTouches[0].clientX
-  } else if (event.type.startsWith('touch')) {
-    clientY = (event as TouchEvent).touches[0].clientY
-    clientX = (event as TouchEvent).touches[0].clientX
-  } else {
-    clientY = (event as MouseEvent).clientY
-    clientX = (event as MouseEvent).clientX
-  }
-  return {
-    clientX,
-    clientY,
-  }
-}
